+++ R/db_writing.R
@@ -4,19 +4,22 @@
 #' into the database. Also adds the cover category for the source.
 #'
 #' @param con connection to the database.
+#' @param schema db schema, defaults to platfrom
 #'
 #' @return nothing
 #' @export
-insert_new_source <- function(con) {
+insert_new_source <- function(con, schema = "platform") {
   SURSfetchR::sql_function_call(con,
                                 "insert_new_source",
-                                as.list(prepare_source_table(con)))
+                                as.list(prepare_source_table(con)),
+                                schema)
   source_id <- UMARaccessR::get_source_code_from_source_name("UMAR", con)
   SURSfetchR::sql_function_call(con,
                                 "insert_new_category",
                                 list(id = 0,
                                      name = "UMAR",
-                                     source_id = source_id[1,1]))
+                                     source_id = source_id[1,1]),
+                                schema)
 }
 
 #' Insert new category
@@ -25,13 +28,14 @@ insert_new_source <- function(con) {
 #'
 #' @param cat_name character name of category - usually full name of author
 #' @param con connection to the database.
-#'
+#' @param schema db schema, defaults to platfrom
 #' @return nothing
 #' @export
-insert_new_category <- function(cat_name, con) {
+insert_new_category <- function(cat_name, con, schema = "platform") {
   SURSfetchR::sql_function_call(con,
                                 "insert_new_category",
-                                as.list(prepare_category_table(cat_name, con)))
+                                as.list(prepare_category_table(cat_name, con)),
+                                schema)
 }
 
 #' Insert new category relationship
@@ -40,11 +44,12 @@ insert_new_category <- function(cat_name, con) {
 #'
 #' @param cat_name character name of category - usually full name of author
 #' @param con connection to the database.
-#'
+#' @param schema db schema, defaults to platfrom
 #' @return nothing
 #' @export
-insert_new_category_relationship <- function(cat_name, con) {
+insert_new_category_relationship <- function(cat_name, con, schema = "platform") {
   SURSfetchR::sql_function_call(con,
                                 "insert_new_category_relationship",
-                                as.list(prepare_category_relationship_table(cat_name, con)))
+                                as.list(prepare_category_relationship_table(cat_name, con)),
+                                schema)
 }