+++ R/globals.R
@@ -1,4 +1,5 @@
-utils::globalVariables(c(".", "name"))
+utils::globalVariables(c(".", "name", "table_name", "interval", "all_same_interval",
+                         "table_code", "gr"))
 #' magrittr pipe
 #'
 #' @importFrom magrittr %>%