+++ tests/testthat/platform/SELECT-af32e3.R
@@ -1,0 +1,2 @@
+structure(list(code = "MZ002"), class = "data.frame", row.names = c(NA, 
+-1L))