+++ tests/testthat/platform/select-9fe446.R
@@ -1,2 +1,2 @@
-structure(list(max = 0L), class = "data.frame", row.names = c(NA, 
+structure(list(max = 1L), class = "data.frame", row.names = c(NA, 
 -1L))