+++ tests/testthat/test-db_writing.R
@@ -1,19 +1,19 @@
 dittodb::with_mock_db({
   con <- DBI::dbConnect(RPostgres::Postgres(),
                         dbname = "platform",
-                        host = "192.168.38.21",
+                        host = "localhost",
                         port = 5432,
-                        user = "majaz",
-                        password = Sys.getenv("PG_MZ_PSW"),
+                        user = "mzaloznik",
+                        password = "kermitit",
                         client_encoding = "utf8")
-  dbExecute(con, "set search_path to platform")
+  dbExecute(con, "set search_path to test_platform")
 
   test_that("prepare functions work", {
     x <- insert_new_source(con)
     expect_equal(dim(x), c(1,1))
     x <- insert_new_category("Maja Zalo\u017enik", con)
     expect_equal(dim(x), c(1,1))
-    x <- insert_new_category_relationship("Maja Zalo\u017enik", con)
+    x <- insert_new_category_relationship("Maja Zalo\u017enik", con, "test_platform")
     expect_equal(dim(x), c(1,1))
   })
 })