+++ tests/testthat/testdata/dittodb_recording_fixtures.R
@@ -53,29 +53,29 @@ library(dittodb)
 # on.exit(dbDisconnect)
 # prepare_category_table("Maja Zalo\u017enik", con)
 # stop_db_capturing()
-
+#
 # start_db_capturing()
 # con <- DBI::dbConnect(RPostgres::Postgres(),
 #                       dbname = "platform",
-#                       host = "192.168.38.21",
+#                       host = "localhost",
 #                       port = 5432,
-#                       user = "majaz",
-#                       password = Sys.getenv("PG_MZ_PSW"),
+#                       user = "mzaloznik",
+#                       password = "kermitit",
 #                       client_encoding = "utf8")
-# dbExecute(con, "set search_path to platform")
+# dbExecute(con, "set search_path to test_platform")
 # on.exit(dbDisconnect)
-# insert_new_category("Maja Zalo\u017enik", con)
+# insert_new_category("Maja Zalo\u017enik", con, "test_platform")
 # stop_db_capturing()
 
 # start_db_capturing()
 # con <- DBI::dbConnect(RPostgres::Postgres(),
 #                       dbname = "platform",
-#                       host = "192.168.38.21",
+#                       host = "localhost",
 #                       port = 5432,
-#                       user = "majaz",
-#                       password = Sys.getenv("PG_MZ_PSW"),
+#                       user = "mzaloznik",
+#                       password = "kermitit",
 #                       client_encoding = "utf8")
-# dbExecute(con, "set search_path to platform")
+# dbExecute(con, "set search_path to test_platform")
 # on.exit(dbDisconnect)
 # prepare_category_relationship_table("Maja Zalo\u017enik", con)
 # stop_db_capturing()
@@ -83,14 +83,26 @@ library(dittodb)
 # start_db_capturing()
 # con <- DBI::dbConnect(RPostgres::Postgres(),
 #                       dbname = "platform",
-#                       host = "192.168.38.21",
+#                       host = "localhost",
 #                       port = 5432,
-#                       user = "majaz",
-#                       password = Sys.getenv("PG_MZ_PSW"),
+#                       user = "mzaloznik",
+#                       password = "kermitit",
 #                       client_encoding = "utf8")
-# dbExecute(con, "set search_path to platform")
+# dbExecute(con, "set search_path to test_platform")
 # on.exit(dbDisconnect)
-# insert_new_category_relationship("Maja Zalo\u017enik", con)
+# insert_new_category_relationship("Maja Zalo\u017enik", con, "test_platform")
 # stop_db_capturing()
 
-
+# start_db_capturing()
+# con <- DBI::dbConnect(RPostgres::Postgres(),
+#                       dbname = "platform",
+#                       host = "localhost",
+#                       port = 5432,
+#                       user = "mzaloznik",
+#                       password = "kermitit",
+#                       client_encoding = "utf8")
+# dbExecute(con, "set search_path to test_platform")
+# on.exit(dbDisconnect)
+# df <- openxlsx::read.xlsx(testthat::test_path("testdata", "struct_tests3.xlsx"), sheet = "M")
+# out <-compute_table_codes(df, con)
+# stop_db_capturing()